requirements.txt
pip==22.3.1
setuptools==65.5.1
wheel==0.38.4
arcface
astropy
tensorflow
numpy
opencv-python
matplotlib
requests
jsonlib
ultralytics
pillow
mtcnn
termcolor
